Simplify Preferences module
remove config.js.
Add preferences_helper.js which handles initialisation and updates.
Move the default preferences values directly into preferences.js.
Note: Might move this into a config json file some time.

We still follow the DIP principles - both high level (most modules) and low level (preferences_helper) modules depend on the abstraction, i.e. the preferences.js module.
