+++ src/api/init.js
@@ -1,9 +1,8 @@
 /// This module initialises everything that is needed for the extension to work.
 /// All modules should be calling this function.
 
+import { initPreferences } from "#root/src/helpers/preferences_helper.js";
 import { initCustomAPIs } from "#root/src/api/providers_custom.js";
-import { initConfig } from "#root/src/api/preferences.js";
-import { initPreferences } from "#root/src/api/preferences.js";
 
 let initialised = false;
 
@@ -14,7 +13,6 @@ export async function init() {
   const start = Date.now();
 
   await initPreferences();
-  initConfig();
   await initCustomAPIs();
 
   console.log(`Extension initialised in ${Date.now() - start}ms`);
