+++ src/api/preferences.js
@@ -4,48 +4,19 @@
 // It is also a small speedup, as we can avoid fetching the preferences multiple times.
 
 import { getPreferenceValues } from "@raycast/api";
-import { Storage } from "#root/src/api/storage.js";
-import { config } from "#root/src/api/config.js";
 
-export const PreferencesStorageKey = "Preferences";
-export const Preferences = getPreferenceValues();
-
-export const loadPreferencesFromStorage = async () => {
-  try {
-    const data = await Storage.read(PreferencesStorageKey);
-    return data ? JSON.parse(data) : {};
-  } catch (error) {
-    console.error("Error loading preferences:", error);
-    return {};
-  }
-};
-
-export const savePreferencesToStorage = async (preferences) => {
-  try {
-    await Storage.write(PreferencesStorageKey, JSON.stringify(preferences));
-  } catch (error) {
-    console.error("Error saving preferences:", error);
-    throw error;
-  }
-};
-
-export const updatePreferences = async (key, value) => {
-  Preferences[key] = value;
-  await savePreferencesToStorage(Preferences);
-};
-
-// Load preferences from storage
-export const initPreferences = async () => {
-  const prefs = await loadPreferencesFromStorage();
-  Object.assign(Preferences, prefs);
-};
-
-// Load config from preferences upon startup
-export const initConfig = () => {
-  for (const key in config) {
-    const value = Preferences[key];
-    if (value !== undefined) {
-      config[key] = value;
-    }
-  }
+export const Preferences = {
+  defaultProvider: "NexraGPT4o",
+  webSearch: "off",
+  proxyURL: "",
+  smartChatNaming: false,
+  autoCheckForUpdates: true,
+  persistentStorage: false,
+  useCursorIcon: true,
+  codeInterpreter: false,
+  devMode: false,
+  GeminiAPIKeys: "",
+  inactiveDuration: 0,
+  defaultLanguage: "English",
+  ...getPreferenceValues(),
 };
